Move form to separate module
